// projects/sphereon-alfresco-blockchain/src/lib/model/verifyNodesResponse.ts
/**
 * Sphereon Alfresco Blockchain Agent
 * This is an API containing functions for blockchain integration with Alfresco.  
 *
 * OpenAPI spec version: 0.1
 * Contact: dev@sphereon.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { VerifyContentResponse } from './verifyContentResponse';


/**
 * Committed context and settings response
 */
export interface VerifyNodesResponse {
    contentResponses: Array<VerifyContentResponse>;
}